moved vprintf to common removed flif.h includes
