+++ common.h
@@ -9,7 +9,6 @@
 #include "maniac/util.h"
 
 #include "image/color_range.h"
-#include "transform/factory.h"
 
 #include "flif_config.h"
 
@@ -23,7 +22,9 @@ extern int64_t pixels_done;
 
 extern const std::vector<std::string> transforms;
 
-
+static int verbosity = 1;
+void increase_verbosity();
+int get_verbosity();
 void v_printf(const int v, const char *format, ...);
 
 typedef SimpleBitChance                         FLIFBitChancePass1;
