+++ transform/transform.h
@@ -5,7 +5,7 @@
 #include "../image/color_range.h"
 #include "../maniac/rac.h"
 #include "../flif_config.h"
-#include "../flif.h"
+#include "../common.h"
 
 
 class Transform {
